+++ scripts/rollback/rollback-2.13.4.sh
@@ -1,0 +1,211 @@
+#!/bin/bash
+# timestamp: 1634792195226
+
+npm dist-tag add @midwayjs/egg-layer@2.13.3 latest
+npm dist-tag add @midwayjs/egg-layer@2.13.4 beta
+npm dist-tag add @midwayjs/express-layer@2.13.3 latest
+npm dist-tag add @midwayjs/express-layer@2.13.4 beta
+npm dist-tag add @midwayjs/koa-layer@2.13.3 latest
+npm dist-tag add @midwayjs/koa-layer@2.13.4 beta
+npm dist-tag add @midwayjs/serverless-app@2.13.3 latest
+npm dist-tag add @midwayjs/serverless-app@2.13.4 beta
+npm dist-tag add @midwayjs/serverless-fc-starter@2.13.3 latest
+npm dist-tag add @midwayjs/serverless-fc-starter@2.13.4 beta
+npm dist-tag add @midwayjs/serverless-fc-trigger@2.13.3 latest
+npm dist-tag add @midwayjs/serverless-fc-trigger@2.13.4 beta
+npm dist-tag add @midwayjs/serverless-scf-trigger@2.13.3 latest
+npm dist-tag add @midwayjs/serverless-scf-trigger@2.13.4 beta
+npm dist-tag add @midwayjs/serverless-vercel-starter@2.13.3 latest
+npm dist-tag add @midwayjs/serverless-vercel-starter@2.13.4 beta
+npm dist-tag add @midwayjs/static-layer@2.13.3 latest
+npm dist-tag add @midwayjs/static-layer@2.13.4 beta
+npm dist-tag add @midwayjs/axios@2.13.3 latest
+npm dist-tag add @midwayjs/axios@2.13.4 beta
+npm dist-tag add @midwayjs/bootstrap@2.13.3 latest
+npm dist-tag add @midwayjs/bootstrap@2.13.4 beta
+npm dist-tag add @midwayjs/cache@2.13.3 latest
+npm dist-tag add @midwayjs/cache@2.13.4 beta
+npm dist-tag add @midwayjs/consul@2.13.3 latest
+npm dist-tag add @midwayjs/consul@2.13.4 beta
+npm dist-tag add @midwayjs/core@2.13.3 latest
+npm dist-tag add @midwayjs/core@2.13.4 beta
+npm dist-tag add @midwayjs/cos@2.13.3 latest
+npm dist-tag add @midwayjs/cos@2.13.4 beta
+npm dist-tag add @midwayjs/faas@2.13.3 latest
+npm dist-tag add @midwayjs/faas@2.13.4 beta
+npm dist-tag add @midwayjs/grpc@2.13.3 latest
+npm dist-tag add @midwayjs/grpc@2.13.4 beta
+npm dist-tag add midway-schedule@2.13.3 latest
+npm dist-tag add midway-schedule@2.13.4 beta
+npm dist-tag add midway@2.13.3 latest
+npm dist-tag add midway@2.13.4 beta
+npm dist-tag add @midwayjs/mock@2.13.3 latest
+npm dist-tag add @midwayjs/mock@2.13.4 beta
+npm dist-tag add @midwayjs/mongoose@2.13.0 latest
+npm dist-tag add @midwayjs/mongoose@2.13.4 beta
+npm dist-tag add @midwayjs/orm@2.13.3 latest
+npm dist-tag add @midwayjs/orm@2.13.4 beta
+npm dist-tag add @midwayjs/oss@2.13.3 latest
+npm dist-tag add @midwayjs/oss@2.13.4 beta
+npm dist-tag add @midwayjs/process-agent@2.13.3 latest
+npm dist-tag add @midwayjs/process-agent@2.13.4 beta
+npm dist-tag add @midwayjs/prometheus-socket-io@2.13.3 latest
+npm dist-tag add @midwayjs/prometheus-socket-io@2.13.4 beta
+npm dist-tag add @midwayjs/prometheus@2.13.3 latest
+npm dist-tag add @midwayjs/prometheus@2.13.4 beta
+npm dist-tag add @midwayjs/rabbitmq@2.13.3 latest
+npm dist-tag add @midwayjs/rabbitmq@2.13.4 beta
+npm dist-tag add @midwayjs/redis@2.13.3 latest
+npm dist-tag add @midwayjs/redis@2.13.4 beta
+npm dist-tag add @midwayjs/sequelize@2.13.3 latest
+npm dist-tag add @midwayjs/sequelize@2.13.4 beta
+npm dist-tag add @midwayjs/socketio@2.13.3 latest
+npm dist-tag add @midwayjs/socketio@2.13.4 beta
+npm dist-tag add @midwayjs/tablestore@2.13.3 latest
+npm dist-tag add @midwayjs/tablestore@2.13.4 beta
+npm dist-tag add @midwayjs/task@2.13.3 latest
+npm dist-tag add @midwayjs/task@2.13.4 beta
+npm dist-tag add @midwayjs/typegoose@2.13.3 latest
+npm dist-tag add @midwayjs/typegoose@2.13.4 beta
+npm dist-tag add @midwayjs/version@2.13.3 latest
+npm dist-tag add @midwayjs/version@2.13.4 beta
+npm dist-tag add @midwayjs/view-ejs@2.13.3 latest
+npm dist-tag add @midwayjs/view-ejs@2.13.4 beta
+npm dist-tag add @midwayjs/view-nunjucks@2.13.3 latest
+npm dist-tag add @midwayjs/view-nunjucks@2.13.4 beta
+npm dist-tag add @midwayjs/view@2.13.3 latest
+npm dist-tag add @midwayjs/view@2.13.4 beta
+npm dist-tag add @midwayjs/express@2.13.3 latest
+npm dist-tag add @midwayjs/express@2.13.4 beta
+npm dist-tag add @midwayjs/koa@2.13.3 latest
+npm dist-tag add @midwayjs/koa@2.13.4 beta
+npm dist-tag add @midwayjs/web@2.13.3 latest
+npm dist-tag add @midwayjs/web@2.13.4 beta
+npm dist-tag add @midwayjs/ws@2.13.3 latest
+npm dist-tag add @midwayjs/ws@2.13.4 beta
+tnpm dist-tag add @midwayjs/egg-layer@2.13.3 latest
+tnpm dist-tag add @midwayjs/egg-layer@2.13.4 beta
+tnpm dist-tag add @midwayjs/express-layer@2.13.3 latest
+tnpm dist-tag add @midwayjs/express-layer@2.13.4 beta
+tnpm dist-tag add @midwayjs/koa-layer@2.13.3 latest
+tnpm dist-tag add @midwayjs/koa-layer@2.13.4 beta
+tnpm dist-tag add @midwayjs/serverless-app@2.13.3 latest
+tnpm dist-tag add @midwayjs/serverless-app@2.13.4 beta
+tnpm dist-tag add @midwayjs/serverless-fc-starter@2.13.3 latest
+tnpm dist-tag add @midwayjs/serverless-fc-starter@2.13.4 beta
+tnpm dist-tag add @midwayjs/serverless-fc-trigger@2.13.3 latest
+tnpm dist-tag add @midwayjs/serverless-fc-trigger@2.13.4 beta
+tnpm dist-tag add @midwayjs/serverless-scf-trigger@2.13.3 latest
+tnpm dist-tag add @midwayjs/serverless-scf-trigger@2.13.4 beta
+tnpm dist-tag add @midwayjs/serverless-vercel-starter@2.13.3 latest
+tnpm dist-tag add @midwayjs/serverless-vercel-starter@2.13.4 beta
+tnpm dist-tag add @midwayjs/static-layer@2.13.3 latest
+tnpm dist-tag add @midwayjs/static-layer@2.13.4 beta
+tnpm dist-tag add @midwayjs/axios@2.13.3 latest
+tnpm dist-tag add @midwayjs/axios@2.13.4 beta
+tnpm dist-tag add @midwayjs/bootstrap@2.13.3 latest
+tnpm dist-tag add @midwayjs/bootstrap@2.13.4 beta
+tnpm dist-tag add @midwayjs/cache@2.13.3 latest
+tnpm dist-tag add @midwayjs/cache@2.13.4 beta
+tnpm dist-tag add @midwayjs/consul@2.13.3 latest
+tnpm dist-tag add @midwayjs/consul@2.13.4 beta
+tnpm dist-tag add @midwayjs/core@2.13.3 latest
+tnpm dist-tag add @midwayjs/core@2.13.4 beta
+tnpm dist-tag add @midwayjs/cos@2.13.3 latest
+tnpm dist-tag add @midwayjs/cos@2.13.4 beta
+tnpm dist-tag add @midwayjs/faas@2.13.3 latest
+tnpm dist-tag add @midwayjs/faas@2.13.4 beta
+tnpm dist-tag add @midwayjs/grpc@2.13.3 latest
+tnpm dist-tag add @midwayjs/grpc@2.13.4 beta
+tnpm dist-tag add midway-schedule@2.13.3 latest
+tnpm dist-tag add midway-schedule@2.13.4 beta
+tnpm dist-tag add midway@2.13.3 latest
+tnpm dist-tag add midway@2.13.4 beta
+tnpm dist-tag add @midwayjs/mock@2.13.3 latest
+tnpm dist-tag add @midwayjs/mock@2.13.4 beta
+tnpm dist-tag add @midwayjs/mongoose@2.13.0 latest
+tnpm dist-tag add @midwayjs/mongoose@2.13.4 beta
+tnpm dist-tag add @midwayjs/orm@2.13.3 latest
+tnpm dist-tag add @midwayjs/orm@2.13.4 beta
+tnpm dist-tag add @midwayjs/oss@2.13.3 latest
+tnpm dist-tag add @midwayjs/oss@2.13.4 beta
+tnpm dist-tag add @midwayjs/process-agent@2.13.3 latest
+tnpm dist-tag add @midwayjs/process-agent@2.13.4 beta
+tnpm dist-tag add @midwayjs/prometheus-socket-io@2.13.3 latest
+tnpm dist-tag add @midwayjs/prometheus-socket-io@2.13.4 beta
+tnpm dist-tag add @midwayjs/prometheus@2.13.3 latest
+tnpm dist-tag add @midwayjs/prometheus@2.13.4 beta
+tnpm dist-tag add @midwayjs/rabbitmq@2.13.3 latest
+tnpm dist-tag add @midwayjs/rabbitmq@2.13.4 beta
+tnpm dist-tag add @midwayjs/redis@2.13.3 latest
+tnpm dist-tag add @midwayjs/redis@2.13.4 beta
+tnpm dist-tag add @midwayjs/sequelize@2.13.3 latest
+tnpm dist-tag add @midwayjs/sequelize@2.13.4 beta
+tnpm dist-tag add @midwayjs/socketio@2.13.3 latest
+tnpm dist-tag add @midwayjs/socketio@2.13.4 beta
+tnpm dist-tag add @midwayjs/tablestore@2.13.3 latest
+tnpm dist-tag add @midwayjs/tablestore@2.13.4 beta
+tnpm dist-tag add @midwayjs/task@2.13.3 latest
+tnpm dist-tag add @midwayjs/task@2.13.4 beta
+tnpm dist-tag add @midwayjs/typegoose@2.13.3 latest
+tnpm dist-tag add @midwayjs/typegoose@2.13.4 beta
+tnpm dist-tag add @midwayjs/version@2.13.3 latest
+tnpm dist-tag add @midwayjs/version@2.13.4 beta
+tnpm dist-tag add @midwayjs/view-ejs@2.13.3 latest
+tnpm dist-tag add @midwayjs/view-ejs@2.13.4 beta
+tnpm dist-tag add @midwayjs/view-nunjucks@2.13.3 latest
+tnpm dist-tag add @midwayjs/view-nunjucks@2.13.4 beta
+tnpm dist-tag add @midwayjs/view@2.13.3 latest
+tnpm dist-tag add @midwayjs/view@2.13.4 beta
+tnpm dist-tag add @midwayjs/express@2.13.3 latest
+tnpm dist-tag add @midwayjs/express@2.13.4 beta
+tnpm dist-tag add @midwayjs/koa@2.13.3 latest
+tnpm dist-tag add @midwayjs/koa@2.13.4 beta
+tnpm dist-tag add @midwayjs/web@2.13.3 latest
+tnpm dist-tag add @midwayjs/web@2.13.4 beta
+tnpm dist-tag add @midwayjs/ws@2.13.3 latest
+tnpm dist-tag add @midwayjs/ws@2.13.4 beta
+
+# Changes:
+
+#  - @midwayjs/egg-layer: 2.13.3 => 2.13.4
+#  - @midwayjs/express-layer: 2.13.3 => 2.13.4
+#  - @midwayjs/koa-layer: 2.13.3 => 2.13.4
+#  - @midwayjs/serverless-app: 2.13.3 => 2.13.4
+#  - @midwayjs/serverless-fc-starter: 2.13.3 => 2.13.4
+#  - @midwayjs/serverless-fc-trigger: 2.13.3 => 2.13.4
+#  - @midwayjs/serverless-scf-trigger: 2.13.3 => 2.13.4
+#  - @midwayjs/serverless-vercel-starter: 2.13.3 => 2.13.4
+#  - @midwayjs/static-layer: 2.13.3 => 2.13.4
+#  - @midwayjs/axios: 2.13.3 => 2.13.4
+#  - @midwayjs/bootstrap: 2.13.3 => 2.13.4
+#  - @midwayjs/cache: 2.13.3 => 2.13.4
+#  - @midwayjs/consul: 2.13.3 => 2.13.4
+#  - @midwayjs/core: 2.13.3 => 2.13.4
+#  - @midwayjs/cos: 2.13.3 => 2.13.4
+#  - @midwayjs/faas: 2.13.3 => 2.13.4
+#  - @midwayjs/grpc: 2.13.3 => 2.13.4
+#  - midway-schedule: 2.13.3 => 2.13.4
+#  - midway: 2.13.3 => 2.13.4
+#  - @midwayjs/mock: 2.13.3 => 2.13.4
+#  - @midwayjs/mongoose: 2.13.0 => 2.13.4
+#  - @midwayjs/orm: 2.13.3 => 2.13.4
+#  - @midwayjs/oss: 2.13.3 => 2.13.4
+#  - @midwayjs/process-agent: 2.13.3 => 2.13.4
+#  - @midwayjs/prometheus-socket-io: 2.13.3 => 2.13.4
+#  - @midwayjs/prometheus: 2.13.3 => 2.13.4
+#  - @midwayjs/rabbitmq: 2.13.3 => 2.13.4
+#  - @midwayjs/redis: 2.13.3 => 2.13.4
+#  - @midwayjs/sequelize: 2.13.3 => 2.13.4
+#  - @midwayjs/socketio: 2.13.3 => 2.13.4
+#  - @midwayjs/tablestore: 2.13.3 => 2.13.4
+#  - @midwayjs/task: 2.13.3 => 2.13.4
+#  - @midwayjs/typegoose: 2.13.3 => 2.13.4
+#  - @midwayjs/version: 2.13.3 => 2.13.4
+#  - @midwayjs/view-ejs: 2.13.3 => 2.13.4
+#  - @midwayjs/view-nunjucks: 2.13.3 => 2.13.4
+#  - @midwayjs/view: 2.13.3 => 2.13.4
+#  - @midwayjs/express: 2.13.3 => 2.13.4
+#  - @midwayjs/koa: 2.13.3 => 2.13.4
+#  - @midwayjs/web: 2.13.3 => 2.13.4
+#  - @midwayjs/ws: 2.13.3 => 2.13.4